feat: generalize UNS1 FMC profile and add FJS732 and TS CJ525 FMC profiles
